create function to copy blocks files instead of hardcoded
